presently-server/requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==2.1.2
django-rest-framework==0.1.0
djangorestframework==3.8.2
idna==2.7
pytz==2018.5
requests==2.19.1
urllib3==1.23
